api: Add LIBNBD_STRICT_AUTO_FLAG control to nbd_set_strict
We've previously documented that nbd_set_strict() can add new bits,
defaulting to on to provide client-side safety, but which can be
overridden when performing specific integration tests against a
server.  Prior to the introduction of libnbd support for extended
headers, a user could manually disable STRICT_FLAGS and
STRICT_COMMANDS to test the response of an older server receiving an
unexpected NBD_CMD_FLAG_PAYLOAD_LEN; but for convenience sake, we
prefer our default for that flag to now track extended headers and
ignore what the user passes in, which removes the ability we had for
that integration test.  Adding a new strictness knob lets the user be
in charge of that bit's contents once again.  The caveat remains that
disabling the strictness bits makes it possible for a client to
violate the NBD protocol which may have undefined results (the
connection may drop, libnbd may hang, ...), so most clients will never
call nbd_set_strict() in the first place.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230803015720.1090260-26-eblake@redhat.com>