info: Simplify shutdown calls
Depending on command line options and server capabilities, nbdinfo has
legitimate reasons to either be in opt mode or fully connected at the
time a handle is ready to close.  Previously, to avoid unwanted error
messages, we had to manually check state to call the correct function
out of nbd_opt_abort or nbd_shutdown that would not fail; but now that
nbd_shutdown covers both modes by default, it is simpler to not have
to worry about it.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230829222319.534640-4-eblake@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>