rust: Use mio::poll instead of requiring epoll
CI shows our async handle fails to build on FreeBSD and MacOS (where
epoll() is not available as a syscall, and therefore not available as
a Rust crate).  We can instead accomplish the same level-probing
effects by doing a freshly-registered zero-timeout poll with mio
(which defers under the hood to epoll on Linux, and kqueue on BSD).

Fixes: 223a9965 ("rust: async: Create an async friendly handle type")
CC: Tage Johansson <tage.j.lists@posteo.net>
Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <3c16be2f2968cd3db7faeff955293bfd28fc6fcc.1693843565.git.frans.tage@gmail.com>