api: Fix nbd_shutdown to work in opt mode
When nbd_set_opt_mode() was introduced, we did not adjust the behavior
of nbd_shutdown(), so as a result it fails if attempted after the
initial connection but before nbd_opt_abort() or nbd_opt_go().  To
avoid spurious error messages, at least nbdinfo has to perform
special-casing on whether it is in opt mode (call nbd_opt_abort) or
connected (call nbd_aio_disconnect or nbd_shutdown) prior to calling
nbd_close.  But since nbd_shutdown() is already documented as being
more than just a synch wrapper around nbd_aio_disconnect(), it is
nicer to have it default to always gracefully shutting the connection,
without having to hard-code the special casing into each application.

I originally thought of adding a new flag (either opt-in or opt-out)
that toggles whether nbd_shutdown tries opt_abort or fails with EINVAL
in opt mode.  But our assumption is that there are few enough apps
that use opt mode that making this change by default is a bug-fix
(converting an error into success), and adding a flag as an
API-visible witness of whether the bug is present is over-engineering.
nbdinfo benefits (see the next patch), but is compiled in tree, so it
doesn't have to work around older behaviors.  And if it turns out some
other app cares about detecting whether things will work, they can
continue to check current mode and call nbd_opt_abort, instead of
relying on the more powerful nbd_shutdown made possible here.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230829222319.534640-3-eblake@redhat.com>
[eblake: retitle and drop flag; treat as a bug fix on by default, per Rich]
Fixes: ba29ce1d73 ("api: Add nbd_set_opt_mode", v1.4.0)
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>