info: Try harder for graceful disconnect from server
There are a number of ways in which gathering information can fail.
But when possible, it makes sense to let the server know that we are
done talking, as it minimizes the likelihood that nbdinfo's error
message will be obscured by an accompanying error message by the
server complaining about an unclean disconnect.

For example, with a one-off qemu temporarily patched as mentioned in
commit 0f8ee8c6 to advertise sizes larger than 2^63, kicking off
'qemu-nbd -f raw -r file &' produces:

pre-patch:

$ ./run nbdinfo --size nbd://localhost
/home/eblake/libnbd/info/.libs/nbdinfo: nbd_get_size: server claims size 9223372036854781440 which does not fit in signed result: Value too large for defined data type
qemu-nbd: option negotiation failed: Failed to read opts magic: Unexpected end-of-file before all data were read

post-patch:
$ ./run nbdinfo --size nbd://localhost
/home/eblake/libnbd/info/.libs/nbdinfo: nbd_get_size: server claims size 9223372036854781440 which does not fit in signed result: Value too large for defined data type

if nbdinfo is run in the same terminal as qemu-nbd.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230922204754.GY7773@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>