rust: Add overview documentation
As with other language bindings, add a libnbd-rust(3) man page which
gives a simple overview of Rust support, a trivial example, and where
to find out more.  Unfortunately we don't have a final location for
the real documentation (eg. on https://docs.rs/) so we have to refer
to the local build.

Most importantly this new page is linked from libnbd(3).

Reviewed-by: Eric Blake <eblake@redhat.com>