rust: Write a custom translator from POD to rustdoc
For more faithful translation of API documentation, write a custom
translator from POD to rustdoc.

The new translator is not complete.  For example it doesn't deal
properly with nested POD tags.  However it does visibly improve the
existing docs and we can work on improving them more in the future.

Reviewed-by: Eric Blake <eblake@redhat.com>