generator: Make sure man pages are rebuilt if examples change
Since the generator is responsible for generating some of the API
documentation (POD files in docs/*.pod), and because some example code
is inserted directly into the generated POD files, we must make the
generator depend on the example code.

This isn't a complete solution since stamp-generator is not actually
built by a simple 'make'.  It should at least ensure that the POD is
regenerated eventually when some other generated file is needed.

A complete solution probably involves running an explicit
'make -C generator stamp-generator' from the docs/ dir.