interop: Test write, flush and zero operations
After a prior commit removed -DNEEDS_TMPFILE we know that we are
always writing to a temporary file that the test program creates.
Therefore we can extend the interoperability tests to add tests of
write, flush and zero operations, as well as testing that the
operations are reflected in the temporary file (after flush).

Note that nbd-server (v3.25) does not advertise flush support so we
cannot test it, nor that changes are reflected through to the
underlying file for this server.  (See previous commit for why)