interop: Use nbd-server FORCEDTLS mode
Commit c833fa1226 ("interop: Add test of nbd-server with TLS") added
the force_tls boolean but commented it out, with no information why it
was not used.  This flag seems to have been around for a long time (at
least since nbd 3.20, 2019) so it seems better to use it.

Updates: commit c833fa1226092fd51b1211fa195a2a3b791a4ee9