interop: Add nbd-server flush flag
This flag should cause the server to advertise and implement flush
(using fsync/fdatasync).  It's unclear why it's not on by default.

However this flag does not actually work at present.  Reading the
source code of nbd-server it seems as if flush cannot be used in the
[generic] section of the configuration, only in the [export] section.
This seems like a bug in nbd-server to me.  As a result of this, the
change currently has no effect but may motivate me to fix nbd-server
one day.