tests: Use wait_for_pidfile instead of open-coded loops
tests/functions.sh.in has a useful function to wait for a PID file to
appear.  Use it more consistently in tests.

Note that some tests still use the open-coded loop.  This is for the
case where we are running the loop inside nbdkit --run '...', where it
is inconvenient to use library functions from tests/functions.sh.in.