ci: Skip certain deadlocking nbd-server tests on Alpine 3.19
Alpine 3.19 tests in the interop/ subdirectory were hanging, while
both Alpine 3.18 and Edge were fine.

The tests which were hanging were these below (normally they should
skip):

  interop/interop-nbd-server
  interop/list-exports-nbd-server

The hanging problem is in fact already known.  It's because older
versions of Alpine (<= 3.19) have an old version of nbd-server which
has known issues.  We had a CI rule already to skip these tests on
Alpine 3.18, but not 3.19.

On Alpine Edge, nbd-server has been updated so it doesn't have the
problem, which is why the test did not hang there (or skip).