info: Try harder to report contents from nbd-server
nbd-server 3.25 blindly reports a size of 0 for NBD_OPT_INFO,
regardless of whether it is exporting a block device or regular file.
While discussion is under way to get that improved for the next
nbd-server release, in the meantime, we can at least rely on the
accurate size after NBD_OPT_GO to report contents more accurately.  A
more invasive patch would be to unconditionally try NBD_OPT_GO if
NBD_OPT_INFO reports a size of 0.

Message-ID: <20240221165451.1733541-1-eblake@redhat.com>
Signed-off-by: Eric Blake <eblake@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>