common/utils: Add vector_reserve_exactly function
This is the same as vector_reserve but it never tries to allocate any
overhead for future expansion.  This is useful if you know ahead of
time that you will not need to expand the vector further.  It's also
useful with fuzzing where we want the bounds of the vector to be tight
so that ASAN can detect buffer over/underflows.