common/utils: Make vector_reserve_page_aligned allocate exactly
This function is only used in one place, in nbdkit's allocator=malloc
with mlock enabled.  Examining that one use the code only wants to
allocate whole pages, so allocating overhead doesn't seem like it is
useful.

However I'm keeping it as a separate commit in case we need to revert
it.