info: Handle failure of call to file
The buffer allocated by getline() for the lineptr argument needs to be
freed even if getline() fails.
The corollary of this is that even in a failure case the pointer is
non-null but points to garbage data.

Properly handle the failure by freeing the pointer and returning NULL to
indicate that failure to the caller.

RWMJ adds:
The getline(3) manual makes this clear:

   "If *lineptr was set to NULL before the call, then the buffer should  be
   freed by the user program even on failure."

Signed-off-by: Thomas Weißschuh <thomas@t-8ch.de>
Signed-off-by: Richard W.M. Jones <rjones@redhat.com>