ocaml: Reimplement NBD.Buffer.t using Bigarray for zero-copy
NBD.Buffer.t is a buffer type that persists across AIO calls, eg. if
you call NBD.aio_pread then you need to pass in a buffer which
persists for longer than the call itself.  Previously this was
implemented by hand in C, but the disadvantage of that is that it
requires a copy to move data between OCaml code and the persistent
buffer.

After discussion with the OCaml community and after making a similar
change to nbdkit, we found that we could avoid the copy by
reimplementing this using the OCaml Bigarray library.  (This is part
of OCaml stdlib so does not require extra dependencies.)

This change is completely backwards compatible, although to get zero
copy you need to change code to read and write to the Bigarray
directly through the .{i} accessor (or other methods).

For example, old code might do:

  let buf = NBD.Buffer.alloc 512 in
  let cookie = NBD.aio_pread nbd buf 0_L in
  ... polling loop until complete ...
  let data = NBD.Buffer.to_string buf in   (* copy to OCaml string *)
  printf "read %c %c %c\n" data.[0] data.[1] data.[2]

New code can access the buffer directly to avoid the copy:

  let buf = NBD.Buffer.alloc 512 in
  let cookie = NBD.aio_pread nbd buf 0_L in
  ... polling loop until complete ...
  printf "read %c %c %c\n" buf.{0} buf.{1} buf.{2}

See:

https://discuss.ocaml.org/t/is-there-a-simple-library-that-wraps-a-c-mallocd-char-buffer/14323
https://gitlab.com/nbdkit/nbdkit/-/commit/25ffb743f5f227ac88002f8764941735d19091c1

Thanks: Nicolas Ojeda Bar
Thanks: Anil Madhavapeddy
Thanks: Simon Cruanes