interop/block-status-64.c: Fix skip path under valgrind
When executing the test under valgrind, with an old version of
qemu-nbd which doesn't support extended headers, this patch was taken
which did not free up the NBD handle.  This resulting in a valgrind
failure (because of the memory leak) instead of skip.  Fix by freeing
the handle along this path.