rust: Make the struct Cookie internal field fully public
Rust warns:

  error: field `0` is never read
    --> src/types.rs:18:19
     |
  18 | pub struct Cookie(pub(crate) u64);
     |            ------ ^^^^^^^^^^^^^^
     |            |
     |            field in this struct

For some reason the field was only visible within the libnbd crate
("pub(crate)").  It was not actually read anywhere in the code or
tests so the warning was technically correct, the best kind of correct.

However external code could read the cookie of an AIO command, so make
it fully public, which also silences the warning.

Fixes: commit b2511d640bc12b0116e6c013a17aefb4e772a058