tests/connect-uri.c: Don't call strlen on each loop iteration
Eric Blake pointed out that strlen could be called on each loop
iteration because the compiler possibly cannot prove that tolower(3)
never returns NUL (and therefore could never change the length of the
string).  This is a simple thing to fix by caching strlen in a
variable.

Fixes: commit e44297f54db6dec9d9c20701d63d5d7e6575d4f3
Thanks: Eric Blake