nbdfuse: Can't use ?tls-certificates or ?tls-psk-file
In order to use nbdfuse with TLS it's necessary to add a call to nbd_set_uri_allow_local_file(), as shown in the below patch.

However, I'm not familiar enough with the nbdfuse use-cases and motivation for the "allow local file" opt-in to know whether this really should be a command-line argument-based opt in.  I'm guessing it should...