lib/crypto.c: Allow CA verification even if h->hostname is not set
Calling gnutls_session_set_verify_cert with the hostname parameter set
to NULL is permitted:
https://www.gnutls.org/manual/html_node/Core-TLS-API.html#gnutls_005fsession_005fset_005fverify_005fcert

It means that the server's hostname in the certificate will not be
verified but we can at least check that the certificate was signed by
the CA.  This allows the CA to be checked even for connections over
Unix domain sockets.

Example:

  $ rm -f /tmp/sock
  $ nbdkit -U /tmp/sock -f --tls=require --tls-certificates=$HOME/d/nbdkit/tests/pki memory 1G &

Before this change:

  $ nbdinfo 'nbds+unix://?socket=/tmp/sock'
  protocol: newstyle-fixed with TLS, using structured packets
  export="":
	export-size: 1073741824 (1G)
	content: data
	uri: nbds+unix:///?socket=/tmp/sock
  [etc]

(works because it never called gnutls_session_set_verify_cert).

After this change:

  $ nbdinfo 'nbds+unix://?socket=/tmp/sock'
  nbdinfo: nbd_connect_uri: gnutls_handshake: Error in the certificate verification. (15/1)

(fails because system CA does not know about nbdkit's certificate
which is signed by the CA from the nbdkit/tests/pki directory)

  $ nbdinfo 'nbds+unix://?socket=/tmp/sock&tls-certificates=/home/rjones/d/nbdkit/tests/pki'
  protocol: newstyle-fixed with TLS, using structured packets
  export="":
	export-size: 1073741824 (1G)
	content: data
	uri: nbds+unix:///?socket=/tmp/sock&tls-certificates=/home/rjones/d/nbdkit/tests/pki
  [etc]

(works because we supplied the correct CA)

Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>