lib: Implement nbd+ssh:// and nbds+ssh:// URIs
I was annoyed that to connect to an NBD server on a remote machine I
had to open up port 10809 on the firewall of that machine.  Since I
already have SSH access, it would be easier to allow tunnelling the
connection over that.

This is a minimal commit which implements NBD tunnelled over SSH.

It is simply a slightly more sophisticated version of this:

  #!/usr/bin/python3
  import nbd
  h = nbd.NBD()
  h.connect_command(["ssh", "remote",
                     "nc", "localhost", "10809"
     # or:           "nc", "-U", "/tmp/sock"
                     ])

I marked this transport as "Experimental" in the documentation so we
can remove it if it causes unexpected problems.  Note also that this
is enabled in URIs by default, so clients which are interested in
sanitising URIs (from untrusted input) will likely need to disable it.