lib: Add nbd_get_subprocess_pid to return h->pid
While debugging an ugly nbdkit / OCaml crash[1] which only happened
when nbdkit was run as a subprocess of libnbd, I needed a way to start
gdb attached to nbdkit.  It's not possible to do this by running
"gdb --args nbdkit -s ..." as a command for various reasons.

My solution to this was (error checking removed for clarity):

  nbd_connect_command (nbd, { "nbdkit", "-s" /* ... */});
  char pid[100];
  snprintf (pid, sizeof pid, "%d", nbd->pid);
  if (fork () > 0) {
    execl ("/usr/bin/gdb", "gdb", "../server/nbdkit", pid, NULL);
    _exit (1);
  }
  while (access ("/tmp/continue", F_OK) != 0) {
    sleep (1);
  }
  /* Test case continues after the file is created. */

In this case I was fishing directly into the private nbd_handle struct
to get the pid field.  However this seems like a genuine (if rare)
case for being able to get the process ID directly.

The documentation is written to be clear that this is only useful for
debugging cases, only works on some platforms, and shouldn't be relied
on more generally.

[1] https://discuss.ocaml.org/t/free-uninitalized-data-when-calling-caml-c-thread-unregister-on-the-main-thread/15198

Reviewed-by: Eric Blake <eblake@redhat.com>