dump: Add --offset for further limiting the dump
We already have --length (aka --limit) for limiting how far the dump
goes; the obvious counterpart is --offset (aka --skip) for controlling
where to start the dump.

I did not bother to try and print an unaligned partial first line so
that subsequent lines can be nicely aligned, but rather the dump
always starts with a full line even when it displays with unusual
offsets.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20250204191131.976692-3-eblake@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>