golang: Replace () with correct argument decl, for GCC 15
GCC 15 (or new stdc) has deprecated function declarations like:

  int f();

We relied on this previously to avoid having to second-guess how
golang / cgo would internally define a function, and to avoid dealing
with golang's const-correctness problems.  However we have to do it
now for GCC 15.

This produces lots of warnings about the const qualifier being
dropped, but there's nothing we can do about that, and hopefully they
remain warnings and not errors.