copy: Fix corrupted hash on incomplete read
When using synchronous read with unknown file size, if the read was
shorter than request size, we updated the hash with the complete buffer,
inserting leftover bytes from the previous read into the hash.

I'm not sure if there is validation for source size and number of blocks
in the blocks vector, so this can generate a corrupted hash silently.

We probably need to validate later that the image size matches the size
of the hashed data.

I could not reproduce a corrupted hash, the issue discovered by reading
the code.