copy: Consider options when zeroing in synch mode
When writing zeros to the destination using synchronous copy we ignored
the values of --destination-is-zero (or --target-is-zero) and
--allocated.

Example commands affected by this issue:

    nbdcopy --synchronous SRC DST
    nbdcopy - DST <src.img

The --destination-is-zero affects only block devices or NBD destination.
When opening a local file we truncate it and set target_is_zero = true.
Ignoring the flag with block device or NBD will write zeros to the
device when the writes are not needed, decreasing performance with large
sparse images.

The --allocated flag affects only NBD destination, since local files are
always treated as target_is_zero. Ignoring the flag would punch holes in
the destination image when we want to keep it allocated.

Fix to consider both flags, so we have the same behavior in synchronous
and asynchronous copy.

Testing --allocated option shows that we always punched holes:

Before:

    $ ./nbdcopy --synchronous --allocated /bench/50g.img \
        -- [ qemu-nbd -t -e 0 -f raw --discard unmap /bench/tmp.img ]

    $ du -sh /bench/tmp.img
    0	/bench/tmp.img

After:

    $ ./nbdcopy --synchronous --allocated /bench/50g.img \
        -- [ qemu-nbd -t -e 0 -f raw --discard unmap /bench/tmp.img ]

    $ du -sh /bench/tmp.img
    51G	/bench/tmp.img

Skipping zeroing has dramatic performance improvement. Tested by copying
50g empty image using NBD destination.

Before:

    $ time ./nbdcopy --synchronous --destination-is-zero /bench/50g.img \
        -- [ qemu-nbd -t -e 0 -f raw /bench/tmp.img ]

    real  0m18.578s
    user  0m2.497s
    sys   0m6.951s

After:

    $ time ./nbdcopy --synchronous --destination-is-zero /bench/50g.img \
        -- [ qemu-nbd -t -e 0 -f raw /bench/tmp.img ]

    real  0m0.174s
    user  0m0.023s
    sys   0m0.151s