copy: Fix file allocation when using --allocated
When using the --allocated option we keep the file allocated when
writing zeros. However since we truncate the file before the copy, and
set target_is_zero, all zeroing calls do nothing since the file already
reads as zeros. This kept the file sparse in all cases.

Fixed by calling new file_allocate_space() if the --allocated option was
set. The helper tries first file_fallocate(), and if the call is not
supported it falls back to writing zeros manually.

Note that we don't use posix_fallocate() since it does not match our use
case. If fallocate() is not supported it tries to keep existing data by
reading all blocks and zeroing allocated blocks, while we need to zero
all blocks on a new sparse file.