copy: Remove output file in a few tests
In some tests which were originally all derived from one
(copy-sparse.sh) we didn't remove the output file before running the
test, and the test only appended to the output file.  This means if
the output file existed before the test was run, then the test could
fail or act strangely, although probably this never happened in
practice.